Remove IGListKit
Renames reused ListKit functionality for OBAListView.
