+++ OBAKit.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -128,15 +128,6 @@
         }
       },
       {
-        "package": "IGListKit",
-        "repositoryURL": "https://github.com/3a4oT/IGListKit.git",
-        "state": {
-          "branch": "spmBrain",
-          "revision": "507264281abd3a72ff87ae9337a6dbf0934b862d",
-          "version": null
-        }
-      },
-      {
         "package": "leveldb",
         "repositoryURL": "https://github.com/firebase/leveldb.git",
         "state": {
