+++ OBAKit/Controls/VisualEffectViewController.swift
@@ -16,7 +16,7 @@ import UIKit
 /// `view` in `viewDidLoad` with Auto Layout.
 ///
 /// - Note: Subviews must be added to `visualEffectView.contentView`.
-public class VisualEffectViewController: UIViewController, HasVisualEffect {
+public class VisualEffectViewController: UIViewController {
 
     public override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
         super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
